+++ Card.java
@@ -1,14 +1,13 @@
-import java.util.List;
-
 public class Card {
+
     private String name;
-    private float metascore;
-    private float userScore;
-    private float numberOfCopies;
-    private float openingMonthIncome;
+    private int metascore;
+    private int userScore;
+    private int numberOfCopies;
+    private int openingMonthIncome;
 
 
-    public Card(String name, float metascore, float userScore, float numberOfCopies, float openingMonthIncome){
+    public Card(String name, int metascore, int userScore, int numberOfCopies, int openingMonthIncome){
         this.name = name;
         this.metascore = metascore;
         this.userScore = userScore;
@@ -21,19 +20,19 @@ public class Card {
         return name;
     }
 
-    public float getMetascore(){
+    public int getMetascore(){
         return metascore;
     }
 
-    public float getUserScore(){
+    public int getUserScore(){
         return userScore;
     }
 
-    public float getNumberOfCopies(){
+    public int getNumberOfCopies(){
         return numberOfCopies;
     }
 
-    public float getOpeningMonthIncome(){
+    public int getOpeningMonthIncome(){
         return openingMonthIncome;
     }
